+++ src/Network/Network/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 --->
 ## Upcoming Release
+* Updated below commands for feature: Custom routes set/remove on Brooklyn Gateway.
+    - Updated New-AzVirtualNetworkGateway: Added optional parameter -CustomRoute to set the address prefixes as custom routes to set on Gateway.
+    - Updated Set-AzVirtualNetworkGateway: Added optional parameter -CustomRoute to set the address prefixes as custom routes to set on Gateway.
 
 ## Version 1.7.0
 * Updated cmdlets with plural nouns to singular, and deprecated plural names.